refactor(dashboard): group the audit forwarder files into their own folder
